CI: Add linux/arm64 Platform To CI
In addition to `linux/amd64`, CI will also build and
ship `linux/arm64` images.

The CI will be relatively slow since it has to run twice the amount
of jobs on the same runners and emulating arm64.
It is possible to speed up this by archiving digests and
uploading them separately.